catalog-service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<25.0
psycopg2-binary>=2.9,<3.0
